+++ storage_utils/tests/test_settings.py
@@ -29,6 +29,7 @@ test_settings = {
 @mock.patch.dict('storage_utils.settings._global_settings', copy.deepcopy(test_settings),
                  clear=True)
 class TestSettings(unittest.TestCase):
+    @mock.patch.dict(os.environ, {}, clear=True)
     def test_initialize_default(self):
         expected_settings = {
             'stor': {},
@@ -43,7 +44,13 @@ class TestSettings(unittest.TestCase):
                 'object_threads': 10,
                 'segment_threads': 10
             },
-            'swift': {},
+            'swift': {
+                'username': '',
+                'password': '',
+                'auth_url': '',
+                'temp_url_key': '',
+                'num_retries': 0
+            },
             'swift:delete': {
                 'object_threads': 10
             },
@@ -67,17 +74,30 @@ class TestSettings(unittest.TestCase):
         settings._initialize()
         self.assertEquals(settings._global_settings, expected_settings)
 
-    def test_initialize_file(self):
+    @mock.patch.dict(os.environ, {}, clear=True)
+    def test_initialize_w_user_file(self):
         expected_settings = {
-            'stor': {
-                'str_val': 'this is a string'
+            'stor': {},
+            's3': {},
+            's3:upload': {
+                'segment_size': 8388608,
+                'object_threads': 10,
+                'segment_threads': 10
             },
-            'something': {
-                'just': 'another value'
+            's3:download': {
+                'segment_size': 8388608,
+                'object_threads': 10,
+                'segment_threads': 10
             },
             'swift': {
-                'num_retries': 5,
-                'fake_secret_key': '7jsdf0983j""SP{}?//'
+                'username': 'fake_user',
+                'password': 'fake_password',
+                'auth_url': '',
+                'temp_url_key': '',
+                'num_retries': 0
+            },
+            'swift:delete': {
+                'object_threads': 10
             },
             'swift:download': {
                 'container_threads': 10,
@@ -97,8 +117,25 @@ class TestSettings(unittest.TestCase):
             }
         }
         filename = os.path.join(os.path.dirname(__file__), 'file_data', 'test.cfg')
-        settings._initialize(filename)
+        with mock.patch('storage_utils.settings.USER_CONFIG_FILE', filename):
+            settings._initialize()
         self.assertEquals(settings._global_settings, expected_settings)
+
+    @mock.patch.dict(os.environ, {'OS_USERNAME': 'test_username'})
+    @mock.patch.dict(os.environ, {'OS_PASSWORD': 'test_password'})
+    @mock.patch.dict(os.environ, {'OS_NUM_RETRIES': '2'})
+    @mock.patch.dict(os.environ, {'OS_AUTH_URL': 'http://test_auth_url.com'})
+    def test_env_vars_loaded(self):
+        print os.environ
+        print settings.get()
+        settings._initialize()
+        print settings.get()
+        print os.environ
+        initial_settings = settings.get()['swift']
+        self.assertEquals(initial_settings['username'], 'test_username')
+        self.assertEquals(initial_settings['password'], 'test_password')
+        self.assertEquals(initial_settings['num_retries'], 2)
+        self.assertEquals(initial_settings['auth_url'], 'http://test_auth_url.com')
 
     def test_get(self):
         self.assertEquals(settings.get(), test_settings)
@@ -139,6 +176,14 @@ class TestSettings(unittest.TestCase):
     def test_update_wo_settings(self):
         settings.update()
         self.assertEquals(settings._global_settings, test_settings)
+
+    @mock.patch.dict('storage_utils.settings._global_settings',
+                     {'foo': 1}, clear=True)
+    def test_update_validation_error(self):
+        with self.assertRaisesRegexp(ValueError, 'not a valid setting'):
+            settings.update({'foo': {'bar': 3}})
+        with self.assertRaisesRegexp(ValueError, 'not a valid setting'):
+            settings.update({'bar': 4})
 
     def test_use_w_settings(self):
         update_settings = {
@@ -182,7 +227,7 @@ class TestSettings(unittest.TestCase):
             self.assertEquals(settings._global_settings, test_settings)
         self.assertEquals(settings._global_settings, test_settings)
 
-    @mock.patch.dict('storage_utils.settings._global_settings', clear=True)
+    @mock.patch.dict('storage_utils.settings._global_settings', {'foo': ''}, clear=True)
     def test_use_nested_w_update(self):
         settings.update({'foo': 0})
         self.assertEquals(settings.get(), {'foo': 0})
@@ -199,7 +244,7 @@ class TestSettings(unittest.TestCase):
         settings.update({'foo': 3})
         self.assertEquals(settings.get(), {'foo': 3})
 
-    @mock.patch.dict('storage_utils.settings._global_settings', clear=True)
+    @mock.patch.dict('storage_utils.settings._global_settings', {'foo': ''}, clear=True)
     def test_use_update_w_error(self):
         with settings.use({'foo': 1}):
             with self.assertRaises(RuntimeError):
@@ -211,7 +256,7 @@ class TestSettings(unittest.TestCase):
             self.assertEquals(settings.get(), {'foo': value})
             time.sleep(.01)
 
-    @mock.patch.dict('storage_utils.settings._global_settings', clear=True)
+    @mock.patch.dict('storage_utils.settings._global_settings', {'foo': ''}, clear=True)
     def test_use_multithreaded(self):
         threads = []
         for i in range(30):