+++ docs/release_notes.rst
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+v9.0.0
+------
+
+* All swift settings use the settings module now (no more swift.update_settings)
+
 v8.5.0
 ------
 
@@ -246,7 +251,7 @@ v4.0.0
 v3.0.0
 ------
 
-* Added `swift.update_settings` function for updating swift module settings. 
+* Added ``swift.update_settings`` function for updating swift module settings. 
   Settings may no longer be changed at module level.
 
 v2.3.0